// frontend/src/api/generated/v1/models/weddingParticipantRoleEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Wedding Management API (Ninja)
 * OpenAPI spec version: 1.0.0
 */

export type WeddingParticipantRoleEnum = typeof WeddingParticipantRoleEnum[keyof typeof WeddingParticipantRoleEnum];


export const WeddingParticipantRoleEnum = {
  BRIDE: 'BRIDE',
  GROOM: 'GROOM',
  FINANCIAL_PAYER: 'FINANCIAL_PAYER',
  LEGAL_REPRESENTATIVE: 'LEGAL_REPRESENTATIVE',
  OTHER: 'OTHER',
} as const;